==> notebooks/3.2-jrb-cma-pavillions.ipynb <==
import sys
import os

project_root = os.path.abspath('..')
if project_root not in sys.path:
    sys.path.append(project_root)
 
%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np

import matplotlib.pyplot as plt

import proyeccion_rdr.features as features
import proyeccion_rdr.produccion.a03_pabellon as pb

pd.set_option('display.max_columns', None)

# Agrega las columnas de poblacion de interes
ANIOS_INTERES = [f"{i}" for i in [2025, 2030, 2040]]

## Obtencion de casos por area de influencia RDR

# Lee los casos de todos los macroprocesos
RUTA_ARCHIVOS = "../data/interim/0.1_casos_teoricos_diagnosticos.xlsx"
_, casos_macroproceso = features.leer_casos_macroprocesos(RUTA_ARCHIVOS)

# Obtiene los casos de CMA
casos_cma = casos_macroproceso.query("tipo_paciente == 'cma'")

## Obtencion de porcentajes de CMA

# Define la ruta donde estan los porcentajes de CMA
RUTA_PLANILLA_PORCENTAJES = (
    "../data/raw/3_incidencias_y_porcentajes_marcoprocesos/incidencias_y_prevalencias_RDR.xlsx"
)

# Solamente lee los diagnosticos y el porcentaje de CMA
datos_cma = pd.read_excel(
    RUTA_PLANILLA_PORCENTAJES,
    usecols=["Diagnostico", "Porcentaje Pacientes CMA", "Horas en CMA"],
)

# Deja los diagnosticos que SI aportan a la cantidad de pabellones de CMA (porcentaje > 0)
datos_cma = datos_cma.query("`Porcentaje Pacientes CMA` > 0")

# Formatea la columna de diagnosticos
datos_cma["Diagnostico"] = datos_cma["Diagnostico"].str.split(" - ").str[0]

# Deja como indice a los diagnosticos
datos_cma = datos_cma.set_index("Diagnostico")

# Indica el tiempo quirurgico por diagnostico promedio (o 75%)
tiempos_cma = datos_cma["Horas en CMA"]

# Multiplica los casos en CMA con sus tiempos quirurgicos
horas_en_cma_por_casos = casos_cma[ANIOS_INTERES].mul(tiempos_cma, axis=0)

# Obtiene la cantidad de horas laborales
horas_laborales = features.calcular_horas_laborales(
    2017, 2040, horas_por_dia_semana=12, incluir_sabado=True, horas_sabado=6
)

# Obtiene cantidad de pabellones de hemodinamia
pabellones_cma = pb.calcular_cantidad_de_pabellones_necesarios(
    horas_en_cma_por_casos, horas_laborales
)

# Obtiene la cantidad de pabellones totales por anio
cantidad_de_pabellones_necesarios = pabellones_cma.sum()

print(f"> Pabellones de CMA 2040: {cantidad_de_pabellones_necesarios['2040']:.2f}")

## Obtener resumen para MINSAL

# Obtiene porcentaje de hospitalizados
porcentaje_cma = pd.read_excel(
    "../data/raw/3_incidencias_y_porcentajes_marcoprocesos/incidencias_y_prevalencias_RDR.xlsx"
)

porcentaje_cma["Diagnostico"] = porcentaje_cma["Diagnostico"].str.split(" - ").str[0]

# Filtra los porcentajes de hospitalizados
porcentaje_cma = porcentaje_cma.set_index("Diagnostico")
porcentaje_cma = porcentaje_cma["Porcentaje Pacientes CMA"]
porcentaje_cma = porcentaje_cma.dropna()
porcentaje_cma = porcentaje_cma[porcentaje_cma > 0]

# Carga el resumen de area de estudio para MINSAL
resumen_area_de_estudio_minsal = pd.read_excel(
    "../data/interim/0.1_casos_teoricos_diagnosticos.xlsx", sheet_name="resumen_total_RDR"
).sort_values("Diagnostico")
resumen_area_de_estudio_minsal["diagnostico_separado"] = (
    resumen_area_de_estudio_minsal["Diagnostico"].str.split(" - ").str[0]
)
resumen_area_de_estudio_minsal = resumen_area_de_estudio_minsal.set_index("diagnostico_separado")

# Sintetiza informacion de pabellon para el 2040
resumen_quirurgicos = pd.DataFrame(
    {
        "porcentaje_cma": porcentaje_cma,
        "casos_cma_2040": casos_cma["2040"],
        "tiempo_quirurgicos": tiempos_cma,
        "horas_cma_2040": horas_en_cma_por_casos["2040"],
        "horas_laborales_2040_pabellon_12_hrs": horas_laborales["2040"],
        "cantidad_de_pabellones_cma": pabellones_cma["2040"],
    }
)

# Une ambos resumenes segun el diagnostico
resumen_minsal = resumen_area_de_estudio_minsal.merge(
    resumen_quirurgicos, how="inner", left_index=True, right_index=True
)

a_guardar = {
    "resumen_MINSAL": resumen_minsal,
    "casos_cma": casos_cma,
    "resumen_duraciones_cma": tiempos_cma,
    "tiempo_utilizado_pabellon": horas_en_cma_por_casos,
    "horas_laborales_por_anio": horas_laborales,
    "pabellones_desg": pabellones_cma,
    "pabellones_totales": cantidad_de_pabellones_necesarios,
}

with pd.ExcelWriter("../data/interim/2.2_estimacion_pabellones_CMA_RDR.xlsx") as file:
    for nombre_hoja, df_a_guardar in a_guardar.items():
        df_a_guardar.to_excel(file, sheet_name=nombre_hoja)
