==> notebooks/4.0-jrb-ambulatory-medical-boxes.ipynb <==
import sys
import os

project_root = os.path.abspath('..')
if project_root not in sys.path:
    sys.path.append(project_root)
 
%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np

import proyeccion_rdr.features as features
from proyeccion_rdr.produccion.a04_ambulatorio import (
    obtener_distribucion_consultas,
    procesar_incidencias,
    leer_consultas_medicas,
    expandir_serie_rendimientos,
)

pd.set_option("display.max_columns", None)

ANIOS_POBLACION = [f"{i}" for i in [2025, 2030, 2040]]

# Lee los casos de todos los macroprocesos
RUTA_ARCHIVOS = "../data/interim/0.1_casos_teoricos_diagnosticos.xlsx"
casos_macroproceso_por_region, casos_macroproceso = features.leer_casos_macroprocesos(
    RUTA_ARCHIVOS
)

# Pone el indice en el DataFrame de los casos
casos_macroproceso = casos_macroproceso.reset_index().set_index(["Diagnostico", "tipo_paciente"])
casos_macroproceso_por_region = casos_macroproceso_por_region.reset_index().set_index(
    ["Diagnostico", "tipo_paciente"]
)

# Define las areas de influencia presenciales
AREAS_PRESENCIALES = [
    "Metropolitana de Santiago",
    "COMUNAS_SIN_SS_EN_RM",
    "SSMC",
    "SSMN",
    "SSMO",
    "SSMOC",
    "SSMS",
    "SSMSO",
    "Acotado por oferta",
]
casos_macroproceso_por_region["es_presencial"] = np.where(
    casos_macroproceso_por_region["Estrato"].isin(AREAS_PRESENCIALES),
    "Box Presencial",
    "Box de Telemedicina",
)

## Objetivo de este cuadernillo

En este cuadernillo se quieren estimar la cantidad de pacientes ambulatorios que asisitrán al INT para el 2040. Este insumo se utilizará para estimar la cantidad de Box Ambulatorios para cada especialidad. Por lo tanto, es necesario hacer los siguientes calculos:

1. Pacientes Ambulatorios
2. Boxes Ambulatorios

RUTA_INCIDENCIAS = (
    "../data/raw/3_incidencias_y_porcentajes_marcoprocesos/incidencias_y_prevalencias_RDR.xlsx"
)

(
    df_incidencias,
    diagnosticos_ingresados,
    casos_por_especialidad_long,
    casos_totales_por_especialidad_y_grupo,
    casos_a_hacerse_cargo_long,
    casos_a_hacerse_cargo_por_especialidad_grupo_y_presencial,
    casos_a_hacerse_cargo_consolidados,
) = procesar_incidencias(
    RUTA_INCIDENCIAS, "consultas_medicas", casos_macroproceso_por_region, ANIOS_POBLACION
)

## Test de Calidad

En este test se quiere saber si todas las trazadoras estan ingresadas en el macroproceso ambulatorio

# Lee la planilla de trazadoras y aisla
trazadoras_totales = pd.read_excel(RUTA_INCIDENCIAS, sheet_name="trazadoras")
trazadoras_totales = set(trazadoras_totales["Diagnostico"].str.split(" - ").str[0].unique())

# diagnosticos_sin_ingresar = trazadoras_totales - diagnosticos_ingresados
# if len(diagnosticos_sin_ingresar) > 0:
#     raise ValueError(f"Falta ingresar las trazadoras: {diagnosticos_sin_ingresar}")

## Lectura de performance historico de ambulatorio

# Lee las consultas medicas
RUTA_AMBULATORIO = "../data/raw/6_ambulatorio/df_procesada_consultas.csv"
consultas_medicas = leer_consultas_medicas(RUTA_AMBULATORIO)

# Obtiene resumen de consultas acumuladas en el periodo
agrupacion_acumulada = ["especialidad_agrupada"]
distribucion_consultas_medicas_acumuladas, consultas_medicas_por_paciente_acumuladas = (
    obtener_distribucion_consultas(consultas_medicas, agrupacion_acumulada)
)

# Aisla el nombre de la columna que tenga el 75% de las consultas
columna_estadistica_cantidad_consultas = distribucion_consultas_medicas_acumuladas.columns[
    distribucion_consultas_medicas_acumuladas.columns.str.contains("75%")
][0]

# Indica cuantas consultas por cada una de las especialidades
cantidad_consultas_medicas_a_ocupar = distribucion_consultas_medicas_acumuladas[
    columna_estadistica_cantidad_consultas
]

# Filtra las especialidades para solo dejar las que estan en la cartera de servicios
cantidad_consultas_medicas_a_ocupar = cantidad_consultas_medicas_a_ocupar[
    cantidad_consultas_medicas_a_ocupar.index.isin(
        casos_a_hacerse_cargo_por_especialidad_grupo_y_presencial.index.get_level_values(0)
    )
]

## Definicion de desempeno para consultas sin produccion

# Define la cantidad de consultas que tendran las especialidades sin un rendimiento historico
consultas_sin_desempeno = pd.Series(
    {
        "CIRUGÍA CARDIOVASCULAR": 1,
        "CIRUGÍA VASCULAR PERIFÉRICA": 1,
        "PSIQUIATRÍA PEDIATRICA Y DE LA ADOLESCENCIA": 1,
    }
)

# Concatena las especilidades que si tienen desempeno con las que no las tienen
cantidad_consultas_medicas_a_ocupar = pd.concat(
    [cantidad_consultas_medicas_a_ocupar, consultas_sin_desempeno]
)

# Indica la cantidad de consultas que tendran la telemedicina
grupos_de_pacientes = [1, 2]
valores_es_presencial = ["Box Presencial", "Box de Telemedicina"]

# Expande la cantidad de consultas segun grupo de pacientes y valores presenciales
cantidad_consultas_medicas_a_ocupar = expandir_serie_rendimientos(
    cantidad_consultas_medicas_a_ocupar, grupos_de_pacientes, valores_es_presencial
)

TIEMPO_CONSULTA = {
    "15 minutos": 0.25,
    "30 minutos": 0.5,
    "45 minutos": 0.75,
    "60 minutos": 1,
    "20 minutos": 0.333,
    "40 minutos": 0.666,
}

rendimientos_reales = pd.Series(
    {
        "ANESTESIOLOGÍA": TIEMPO_CONSULTA["20 minutos"],
        "CARDIOLOGÍA": TIEMPO_CONSULTA["30 minutos"],
        "CIRUGÍA PEDIÁTRICA": TIEMPO_CONSULTA["30 minutos"],
        "CIRUGÍA PLÁSTICA Y REPARADORA": TIEMPO_CONSULTA["30 minutos"],
        "CIRUGÍA Y TRAUMATOLOGÍA BUCO MAXILOFACIAL": TIEMPO_CONSULTA["30 minutos"],
        "DERMATOLOGÍA": TIEMPO_CONSULTA["20 minutos"],
        "ENDOCRINOLOGÍA PEDIÁTRICA": TIEMPO_CONSULTA["40 minutos"],
        "ENFERMEDADES RESPIRATORIA PEDIÁTRICAS": TIEMPO_CONSULTA["30 minutos"],
        "GASTROENTEROLOGÍA PEDIÁTRICA": TIEMPO_CONSULTA["40 minutos"],
        "GENÉTICA CLÍNICA": TIEMPO_CONSULTA["60 minutos"],
        "GINECOLOGÍA PEDIÁTRICA Y DE LA ADOLESCENCIA": TIEMPO_CONSULTA["30 minutos"],
        "HEMATOLOGÍA": TIEMPO_CONSULTA["40 minutos"],
        "INFECTOLOGÍA": TIEMPO_CONSULTA["40 minutos"],
        "INMUNOLOGÍA": TIEMPO_CONSULTA["40 minutos"],
        "NEFROLOGÍA PEDIÁTRICA": TIEMPO_CONSULTA["40 minutos"],
        "NEUROCIRUGÍA": TIEMPO_CONSULTA["30 minutos"],
        "NEUROLOGÍA PEDIATRICA": TIEMPO_CONSULTA["45 minutos"],
        "OFTALMOLOGÍA": TIEMPO_CONSULTA["20 minutos"],
        "OTORRINOLARINGOLOGÍA": TIEMPO_CONSULTA["20 minutos"],
        "PEDIATRÍA": TIEMPO_CONSULTA["30 minutos"],
        "REUMATOLOGÍA": TIEMPO_CONSULTA["40 minutos"],
        "TRAUMATOLOGÍA Y ORTOPEDIA": TIEMPO_CONSULTA["30 minutos"],
        "UROLOGÍA": TIEMPO_CONSULTA["30 minutos"],
        "CIRUGÍA CARDIOVASCULAR": TIEMPO_CONSULTA["30 minutos"],
        "CIRUGÍA VASCULAR PERIFÉRICA": TIEMPO_CONSULTA["30 minutos"],
        "PSIQUIATRÍA PEDIATRICA Y DE LA ADOLESCENCIA": TIEMPO_CONSULTA["45 minutos"],
    }
)

# Expande los rendimientos
rendimientos_reales = expandir_serie_rendimientos(
    rendimientos_reales, grupos_de_pacientes, valores_es_presencial
)

## Estimacion de Consultas al 2040

# Obtiene la cantidad de horas laborales
horas_laborales = features.calcular_horas_laborales(2017, 2040, 12)

# Multiplica los casos de area de influencia
consultas_proyectadas = casos_a_hacerse_cargo_por_especialidad_grupo_y_presencial.mul(
    cantidad_consultas_medicas_a_ocupar, axis=0
).dropna(axis=0, how="all")

# Multiplica la cantidad de consultas por el tiempo que requerira cada consulta
tiempo_consultas = consultas_proyectadas.mul(rendimientos_reales, axis=0).dropna(axis=0, how="all")

# Divide la cantidad de consultas por especialidad, por el rendimiento de cada box
boxes_proyectados = tiempo_consultas.div(horas_laborales, axis=1)

# Obtiene el resumen de Boxes por estamnento
boxes_consolidados = (
    boxes_proyectados.reset_index()
    .groupby("ESTAMENTO/ESPECIALIDAD")
    .sum()["2040"]
    .sort_values(ascending=False)
)

# Agrega el total
boxes_consolidados["Total"] = boxes_consolidados.sum()

display(boxes_consolidados)

# Obtiene la cantidad de boxes medicos y de telemedicina
boxes_presencial_y_tele = (
    boxes_proyectados.reset_index().groupby("es_presencial")[ANIOS_POBLACION].sum()
)

boxes_presencial_y_tele

# Define columnas que se utilizar para generar resumen MINSAL
cols_indice = ["ESTAMENTO/ESPECIALIDAD", "Grupos de Pacientes", "es_presencial"]

# Genera un DataFrame de las trazadoras con presencialidad
df_incidencias_resumen = df_incidencias.copy()
df_incidencias_resumen["es_presencial"] = "Box Presencial, Box de Telemedicina"
df_incidencias_resumen["es_presencial"] = df_incidencias_resumen["es_presencial"].str.split(", ")
df_incidencias_resumen = df_incidencias_resumen.explode("es_presencial")
df_incidencias_resumen = df_incidencias_resumen.set_index(cols_indice)

## Resumen MINSAL

# Obtiene el resumen MINSAL
resumen_MINSAL = pd.DataFrame(
    {
        "tipo_paciente": df_incidencias_resumen["tipo_paciente"],
        "Diagnostico": df_incidencias_resumen["Diagnostico"],
        "Areas de Influencia a atender presencial": str(AREAS_PRESENCIALES),
        "numero_de_pacientes_totales_2025": casos_totales_por_especialidad_y_grupo["2025"],
        "numero_de_pacientes_totales_2030": casos_totales_por_especialidad_y_grupo["2030"],
        "numero_de_pacientes_totales_2040": casos_totales_por_especialidad_y_grupo["2040"],
        "% de los pacientes a atender": df_incidencias_resumen["% de los pacientes a atender"],
        "explicacion_pacientes": df_incidencias_resumen["Explicación Pacientes"],
        "casos_a_hacerse_cargo_2040": casos_a_hacerse_cargo_por_especialidad_grupo_y_presencial[
            "2040"
        ],
        "consultas_por_paciente_75%": cantidad_consultas_medicas_a_ocupar,
        "consultas_proyectadas_2025": consultas_proyectadas["2025"],
        "consultas_proyectadas_2030": consultas_proyectadas["2030"],
        "consultas_proyectadas_2040": consultas_proyectadas["2040"],
        "horas_por_consultas": rendimientos_reales,
        "horas_consultas_2040": tiempo_consultas["2040"],
        "horas_laborales_2040": horas_laborales["2040"],
        "boxes_proyectados_2040": boxes_proyectados["2040"],
    }
)

resumen_MINSAL = resumen_MINSAL.dropna().reset_index(level=[1, 2])

a_guardar = {
    "resumen_MINSAL": resumen_MINSAL,
    "boxes_consolidados": boxes_consolidados,
    "trazadoras_ambulatorio": df_incidencias,
    "casos_macroprocesos_por_region": casos_macroproceso_por_region.reset_index(),
    "casos_por_esp_long": casos_por_especialidad_long.reset_index(),
    "casos_a_hacerse_cargo_long": casos_a_hacerse_cargo_long,
    "casos_por_esp_grupo_y_pres": casos_a_hacerse_cargo_por_especialidad_grupo_y_presencial,
    "casos_consolidados": casos_a_hacerse_cargo_consolidados,
    "metricas_historicas_por_espec": distribucion_consultas_medicas_acumuladas,
    "rendimientos_utilizados": rendimientos_reales,
    "consultas_medicas_proyectadas": consultas_proyectadas,
    "tiempo_por_especialidad": tiempo_consultas,
    "horas_laborales": horas_laborales,
    "boxes_medicos_proyectados": boxes_proyectados,
    "boxes_presenciales_y_tele": boxes_presencial_y_tele,
}

with pd.ExcelWriter("../data/interim/3.0_estimacion_boxes_medicos_RDR.xlsx") as file:
    for nombre_hoja, df_a_guardar in a_guardar.items():
        df_a_guardar.to_excel(file, sheet_name=nombre_hoja)
