==> notebooks/5.0-jrb-support-units-percentajes.ipynb <==
import sys
import os

project_root = os.path.abspath('..')
if project_root not in sys.path:
    sys.path.append(project_root)
 
%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np

import proyeccion_rdr.features as features
import proyeccion_rdr.produccion.a05_unidades_de_apoyo as ua
from proyeccion_rdr.produccion.a04_ambulatorio import (
    leer_consultas_medicas
)


pd.set_option("display.max_columns", None)

ANIOS_POBLACION = [str(i) for i in range(2017, 2036)]

## Leer insumos de Apoyo Diagnostico

# Lee la proyeccion de consultas medicas
RUTA_CONSULTAS = "../data/interim/3.0_estimacion_boxes_medicos_RDR.xlsx"
consultas_medicas_proyectadas = features.leer_proyeccion_consultas_medicas_especialidad(
    RUTA_CONSULTAS, ANIOS_POBLACION
)

# Lee la cantidad de DCO proyectados
RUTA_DCO = "../data/interim/1.0_estimacion_atencion_cerrada_RDR.xlsx"
dco_proyectados = features.leer_dco_proyectados(RUTA_DCO, ANIOS_POBLACION)

# Define las rutas de los archivos
ruta_raw = "../data/raw/"

# Define las rutas de la base de CAE y GRD
ruta_cae = ruta_raw + "6_ambulatorio/df_procesada_consultas.csv"
ruta_grd = ruta_raw + "7_egresos_grd/df_procesada_grd.csv"
ruta_urgencias = ruta_raw + "8_urgencias/df_procesada_urgencias.csv"

# Define las rutas de los archivos de apoyo diagnostico
ruta_apoyo = ruta_raw + "9_unidades_de_apoyo/"

ruta_laboratorio = ruta_apoyo + "df_procesada_laboratorio_clinico.csv"
ruta_farmacia = ruta_farmacia_amb = ruta_apoyo + "df_procesada_farmacia.csv"
ruta_imagenologia = ruta_apoyo + "df_procesada_imagenologia.csv"
ruta_umt = ruta_apoyo + "df_procesada_umt.csv"

# Lee CAE y GRD
consultas_medicas = leer_consultas_medicas(ruta_cae)
df_grd = ua.leer_grd(ruta_grd)

# Lee las unidades de apoyo
df_laboratorio, df_laboratorio_hosp, df_laboratorio_amb = ua.leer_laboratorio(ruta_laboratorio)
df_farmacia, df_farmacia_hosp, df_farmacia_amb = ua.leer_farmacia(ruta_farmacia)

# Define la agrupacion a calcular
AGRUPACION_EXAMENES = ["ano", "id_paciente"]

# Obtiene los dco totales y consultas por pacientes por anio
consultas_por_pacientes = consultas_medicas.groupby(AGRUPACION_EXAMENES)["fecha_atencion"].count()
dco_por_pacientes = df_grd.groupby(AGRUPACION_EXAMENES)["estancia"].sum()

# Obtiene los examanes por paciente de laboratorio
examenes_ppct_lab_hosp = df_laboratorio_hosp.groupby(AGRUPACION_EXAMENES)["examen"].count()
examenes_ppct_lab_amb = df_laboratorio_amb.groupby(AGRUPACION_EXAMENES)["examen"].count()

# Obtiene las prescripciones por pacientes de farmacia
prescr_ppct_farmacia_hosp = df_farmacia_hosp.groupby(AGRUPACION_EXAMENES)["nombre_articulo"].count()
prescr_ppct_farmacia_amb = df_farmacia_amb.groupby(AGRUPACION_EXAMENES)["nombre_articulo"].count()

# Obtiene los examenes por DCO y consultas para laboratorio
_, examenes_por_dco_lab = ua.calcular_examenes_por_atencion(
    dco_por_pacientes, examenes_ppct_lab_hosp, "estancia", "examen"
)
_, examenes_por_consulta_lab = ua.calcular_examenes_por_atencion(
    consultas_por_pacientes, examenes_ppct_lab_amb, "fecha_atencion", "examen"
)

# Obtiene los examenes por DCO y consultas para laboratorio
_, prescr_por_dco = ua.calcular_examenes_por_atencion(
    dco_por_pacientes, prescr_ppct_farmacia_hosp, "estancia", "nombre_articulo"
)
_, prescr_por_consulta = ua.calcular_examenes_por_atencion(
    consultas_por_pacientes, prescr_ppct_farmacia_amb, "fecha_atencion", "nombre_articulo"
)

## Laboratorio Clinico

En este apartado se quiere saber cuantos examenes de Laboratorio Clinico se haran

print("Laboratorio Clinico - AC - Examenes por DCO")
display(examenes_por_dco_lab)

print("Laboratorio Clinico - AA - Examenes por CME")
display(examenes_por_consulta_lab)

Como se ve, la cantidad de examenes por CME es muy baja, por lo que se ocupara el estandar MINSAL.

# Define constantes
CANTIDAD_EXAMENES_AC_A_HACER = examenes_por_dco_lab["75%"].max()
CANTIDAD_EXAMENES_AA_A_HACER = 3 # Estandar MINSAL
CANTIDAD_EXAMENES_URG_A_HACER = 2.5 # Estandar MINSAL
metricas_para_proyectar_laboratorio = pd.DataFrame(
    {
        "examenes_por_dco": [CANTIDAD_EXAMENES_AC_A_HACER],
        "examenes_por_consulta": [CANTIDAD_EXAMENES_AA_A_HACER],
    }
)

# Obtiene le cantidad de examenes de laboratorio requeridos para atencion cerrada
examenes_laboratorio_ac = dco_proyectados * CANTIDAD_EXAMENES_AC_A_HACER

# Obtiene la cantidad de examenes de laboratorio requeridos para atencion abierta
examanes_laboratorio_aa = consultas_medicas_proyectadas * CANTIDAD_EXAMENES_AA_A_HACER

# Suma la cantidad de examenes totales
examenes_laboratorio_proyectados = examanes_laboratorio_aa + examenes_laboratorio_ac

# Muestra la produccion actual
produccion_actual_laboratorio = (
    df_laboratorio_amb.groupby("ano").size() + df_laboratorio_hosp.groupby("ano").size()
)

display(produccion_actual_laboratorio)
display(round(examenes_laboratorio_proyectados))

produccion_promedio_laboratorio = produccion_actual_laboratorio[
    produccion_actual_laboratorio.index != 2025
].mean()
crecimiento_laboratorio = examenes_laboratorio_proyectados["2035"] / produccion_promedio_laboratorio
print(f"Aumento de Laboratorio: {crecimiento_laboratorio:.2f} veces")

## Farmacia

print("Farmacia - AC - Prescripciones por DCO")
display(prescr_por_dco)

print("Farmacia - AA - Prescripciones por CME")
display(prescr_por_consulta)

# Define constantes
PRESCRIPCIONES_POR_DCO = prescr_por_dco["75%"].max()
PRESCRIPCIONES_POR_CME = 3 # Estandar MINSAL
metricas_para_proyectar_farmacia = pd.DataFrame(
    {
        "prescripciones_por_dco": [PRESCRIPCIONES_POR_DCO],
        "prescripciones_por_consulta": [PRESCRIPCIONES_POR_CME],
    }
)

# # Obtiene la cantidad de prescripciones
prescripciones_hosp = dco_proyectados * PRESCRIPCIONES_POR_DCO
prescripciones_amb = consultas_medicas_proyectadas * PRESCRIPCIONES_POR_CME

# # Obtiene el total de prescripciones
prescripciones_proyectadas = prescripciones_hosp + prescripciones_amb

produccion_actual_farmacia = (
    df_farmacia_amb.groupby("ano").size() + df_farmacia_hosp.groupby("ano").size()
)
display(produccion_actual_farmacia)

display(round(prescripciones_proyectadas))

produccion_promedio_farmacia = produccion_actual_farmacia.mean()
crecimiento_farmacia = prescripciones_proyectadas["2035"] / produccion_promedio_farmacia
print(f"Aumento de farmacia: {crecimiento_farmacia:.2f} veces")

## Guardar archivos

# Consolida los casos utilizados para proyectar las unidades de apoyo
resumen_casos_apoyo_dco_cme = pd.DataFrame(
    {
        "DCO Proyectados": dco_proyectados,
        "CME Proyectados": consultas_medicas_proyectadas,
    }
).transpose()
resumen_casos_apoyo_dco_cme["utilizado_en"] = "Laboratorio Clinico - Farmacia"

# Une los casos de apoyo
resumen_casos_apoyo = pd.concat(
    [
        resumen_casos_apoyo_dco_cme,
    ]
).reset_index(names="Trazadoras")

a_guardar = {
    "DCO, CME y Casos Proyecciones": {
        "Casos Utilizados para Proyectar": resumen_casos_apoyo,
    },
    "Laboratorio Clinico": {
        "Exámenes por DCO Históricos": examenes_por_dco_lab.reset_index(),
        "Exámenes por CME Históricos": examenes_por_consulta_lab.reset_index(),
        "Métricas seleccionadas para proyectar": metricas_para_proyectar_laboratorio.reset_index(),
        "Exámenes Proyectados - AC ": examenes_laboratorio_ac.reset_index(),
        "Exámenes Proyectados - AA ": examanes_laboratorio_aa.reset_index(),
        "Exámenes Proyectados - Totales": examenes_laboratorio_proyectados.reset_index(),
    },
    "Farmacia": {
        "Prescripciones por DCO Históricas": prescr_por_dco.reset_index(),
        "Prescripciones por CME Históricas": prescr_por_consulta.reset_index(),
        "Métricas seleccionadas para proyectar": metricas_para_proyectar_farmacia.reset_index(),
        "Prescripciones Proyectadas - AC": prescripciones_hosp.reset_index(),
        "Prescripciones Proyectadas - AA": prescripciones_amb.reset_index(),
        "Prescripciones Proyectada - Total": prescripciones_proyectadas.reset_index(),
    },
}

# Guarda archivos de Apoyo
ruta_a_guardar = "../data/interim/4.0_estimacion_unidades_de_apoyo_RDR.xlsx"
ua.guardar_dict_en_excel(ruta_a_guardar, a_guardar, espacio_filas=3)
